+++ src/main/java/com/skyflow/common/utils/Helpers.java
@@ -103,6 +103,38 @@ public final class Helpers {
         return postRequestInput;
 
     }
+    public static JSONObject constructQueryRequest(QueryRecordInput queryInput, QueryOptions options)
+            throws SkyflowException {
+        if (queryInput == null) {
+            throw new SkyflowException(ErrorCode.EmptyRecords);
+        }
+        JSONObject postRequestInput = new JSONObject();
+            if (queryInput.getQuery() == null||queryInput.getQuery().trim().isEmpty()) {
+                throw new SkyflowException(ErrorCode.InvalidQuery);
+            }
+        postRequestInput.put("query", queryInput.getQuery());
+        return postRequestInput;
+    }
+
+    public static JSONObject constructQueryErrorObject(SkyflowException skyflowException) {
+        JSONObject finalResponseError = new JSONObject();
+        try {
+            JSONObject errorObject = (JSONObject) ((JSONObject) new JSONParser().parse(skyflowException.getMessage())).get("error");
+            if (errorObject != null) {
+                JSONObject responseError = new JSONObject();
+                responseError.put("code", errorObject.get("http_code"));
+                responseError.put("description", errorObject.get("message"));
+                finalResponseError.put("error", responseError);
+            }
+        } catch (ParseException e) {
+            JSONObject responseError = new JSONObject();
+            responseError.put("code", skyflowException.getCode());
+            responseError.put("description", skyflowException.getMessage());
+            finalResponseError.put("error", responseError);
+
+        }
+        return finalResponseError;
+    }
 
     public static StringBuilder constructGetByIdRequestURLParams(GetByIdRecordInput record) {
         StringBuilder paramsList = new StringBuilder();