+++ src/main/java/com/skyflow/entities/DeleteInput.java
@@ -1,0 +1,13 @@
+package com.skyflow.entities;
+
+public class DeleteInput {
+    private DeleteRecordInput[] records;
+
+    public DeleteRecordInput[] getRecords() {
+        return records;
+    }
+
+    public void setRecords(DeleteRecordInput[] records) {
+        this.records = records;
+    }
+}