+++ src/main/java/com/skyflow/errors/ErrorCode.java
@@ -20,12 +20,16 @@ public enum ErrorCode {
     BearerThrownException(400, "getBearer() thrown exception"),
     EmptyRecords(400, "Records cannot be empty"),
     InvalidTable(400, "Table name is missing"),
+    InvalidId(400, "Skyflow id is missing"),
     InvalidFields(400, "Fields are missing"),
+    InvalidQueryInput(400, "Invalid query input"),
+    InvalidQuery(400, "Query is missing"),
     InvalidSkyflowId(400, "Skyflow id are missing"),
     InvalidToken(400, "Token is empty"),
     InvalidDetokenizeInput(400, "Invalid Detokenize Input"),
     InvalidInsertInput(400, "Invalid insert input"),
     InvalidUpdateInput(400, "Invalid update input"),
+    InvalidDeleteInput(400, "Invalid delete input"),
     InvalidGetByIdInput(400, "Invalid getById input"),
     InvalidGetInput(400, "Invalid get input"),
     MissingIdAndColumnName(400, "Provide either Ids or column name to get records."),
@@ -68,7 +72,7 @@ public enum ErrorCode {
         return code;
     }
 
-    public String getDescription() {
+    public String  getDescription() {
         return description;
     }
 }