+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [1.11.0] - 2023-09-01
+### Added
+- `query` vault API
+
+## [1.10.0] - 2023-08-09
+- Added `delete` vault API support. 
 ## [1.9.0] - 2023-06-08
 ### Added
 - `redaction` key for detokenize method for column group support.